atlas-mvp/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
streamlit==1.34.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
sentence-transformers==2.6.1
numpy==1.26.4
ollama==0.1.6
openai==1.14.2
PyYAML==6.0.1
cairosvg==2.7.1
Pillow==10.3.0
SpeechRecognition==3.10.0
pyttsx3==2.90
python-multipart==0.0.9
pre-commit==3.7.0
black==24.3.0
isort==5.13.2
flake8==7.0.0
pytest==8.1.1
